Address Copilot review feedback on LexEntry locking
- Add rubocop:disable for obtain_lock's Naming/PredicateMethod violation
- Fix links_controller missing lex_entry_to_lock override (@entry vs @lex_entry)
- Guard against nil locked_by_user when building lock error message
- Use subquery (select) instead of pluck+where for locked entry exclusion; eager-load :locked_by_user for my_locked_entries to avoid N+1
- Fix migration column type: bigint → integer to match users.id and ingestibles convention
- Add spec for "same user within 10s does not refresh locked_at" edge case

+++ db/migrate/20260526225222_add_lock_to_lex_entries.rb
@@ -5,7 +5,7 @@ class AddLockToLexEntries < ActiveRecord::Migration[8.1]
   def change
     change_table :lex_entries, bulk: true do |t|
       t.datetime :locked_at
-      t.bigint :locked_by_user_id
+      t.integer :locked_by_user_id
       t.index :locked_by_user_id
     end
   end
